upgrading kawala to junit 4.8.2
git-svn-id: http://kawala.googlecode.com/svn/trunk@304 d744d98c-a4b2-11de-97f0-97815e7b2756

+++ srctest/com/kaching/platform/converters/FiniteConverterTest.java
@@ -29,8 +29,8 @@ public class FiniteConverterTest {
 
   @Test
   public void fromAndTo() throws Exception {
-    assertEquals(1, converter.fromString("a"));
-    assertEquals(2, converter.fromString("b"));
+    assertEquals((Integer) 1, converter.fromString("a"));
+    assertEquals((Integer) 2, converter.fromString("b"));
     assertEquals("a", converter.toString(1));
     assertEquals("b", converter.toString(2));
   }
